Graphical Oracle JRE installer for Firefox flatpak
Oracle JRE needs a cookie for accepting the license, therefore
it cannot be installed by extra_data in flatpak. That's why
this utility was created. It is going to be run on flatpak
start when JRE will be unavailable. It will install JRE in
the user directory, which is shared between host and flatpak.